+++ apis/keda/v1alpha1/triggerauthentication_types.go
@@ -249,6 +249,9 @@ type Credential struct {
 	Token string `json:"token,omitempty"`
 
 	// +optional
+	TokenFrom *ValueFromSecret `json:"tokenFrom,omitempty"`
+
+	// +optional
 	ServiceAccount string `json:"serviceAccount,omitempty"`
 
 	// +optional